Change set for the Philemon release (Alpha Build 17)
This version of the Essence# Standard Library is, and all subsequent
versions will be,  incompatible with any Essence# run time system prior
to the Philemon release (Alpha Build 17).  However, the reverse is not
true: The Philemon release of the Essence# run time system can be used
with any class library that uses the old class library format.

Please see the release notes on the CodePlex Download page for further
details.

This version also includes new code: Essence# classes that represent all
the core .Net classes for dealing with the file system.